# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType];

export const CategoryType = {
  text: "text",
  image: "image",
} as const;

export interface Category {
  id: number;
  name: string;
  description?: string | null;
  type: CategoryType;
  enabled: boolean;
  itemCount: number;
  createdAt: string;
}

export type CreateCategoryBodyType =
  (typeof CreateCategoryBodyType)[keyof typeof CreateCategoryBodyType];

export const CreateCategoryBodyType = {
  text: "text",
  image: "image",
} as const;

export interface CreateCategoryBody {
  name: string;
  description?: string | null;
  type: CreateCategoryBodyType;
  enabled?: boolean;
}

export type UpdateCategoryBodyType =
  (typeof UpdateCategoryBodyType)[keyof typeof UpdateCategoryBodyType];

export const UpdateCategoryBodyType = {
  text: "text",
  image: "image",
} as const;

export interface UpdateCategoryBody {
  name?: string;
  description?: string | null;
  type?: UpdateCategoryBodyType;
  enabled?: boolean;
}

export interface CategoryItem {
  id: number;
  categoryId: number;
  itemText: string;
  imageUrl?: string | null;
  createdAt: string;
}

export interface UploadResult {
  imported: number;
  skipped: number;
  errors: string[];
  replaced: boolean;
}

export interface AdminVerifyBody {
  password: string;
}

export interface AdminVerifyResult {
  valid: boolean;
}

export interface CreateRoomBody {
  hostName: string;
  categoryId?: number | null;
  turnDuration?: number | null;
  roundCount?: number | null;
}

export interface JoinRoomBody {
  playerName: string;
}

export interface Player {
  id: number;
  name: string;
  isHost: boolean;
  score: number;
  connected: boolean;
}

export type RoomWithPlayersStatus =
  (typeof RoomWithPlayersStatus)[keyof typeof RoomWithPlayersStatus];

export const RoomWithPlayersStatus = {
  waiting: "waiting",
  playing: "playing",
  finished: "finished",
} as const;

export interface RoomWithPlayers {
  id: number;
  code: string;
  status: RoomWithPlayersStatus;
  categoryId?: number | null;
  categoryName?: string | null;
  turnDuration: number;
  roundCount: number;
  players: Player[];
  createdAt: string;
}

export interface JoinRoomResult {
  room: RoomWithPlayers;
  playerId: number;
  playerName: string;
}

export type AdminDeleteCategory200 = {
  success: boolean;
};

export type AdminUploadItemsBody = {
  file: Blob;
};