# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all active categories
 */
export const ListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  type: zod.enum(["text", "image"]),
  enabled: zod.boolean(),
  itemCount: zod.number(),
  createdAt: zod.string(),
});
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary List all categories (admin)
 */
export const AdminListCategoriesHeader = zod.object({
  "x-admin-password": zod.string(),
});

export const AdminListCategoriesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  type: zod.enum(["text", "image"]),
  enabled: zod.boolean(),
  itemCount: zod.number(),
  createdAt: zod.string(),
});
export const AdminListCategoriesResponse = zod.array(
  AdminListCategoriesResponseItem,
);

/**
 * @summary Create a category
 */
export const AdminCreateCategoryHeader = zod.object({
  "x-admin-password": zod.string(),
});

export const AdminCreateCategoryBody = zod.object({
  name: zod.string(),
  description: zod.string().nullish(),
  type: zod.enum(["text", "image"]),
  enabled: zod.boolean().optional(),
});

/**
 * @summary Update a category
 */
export const AdminUpdateCategoryParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminUpdateCategoryHeader = zod.object({
  "x-admin-password": zod.string(),
});

export const AdminUpdateCategoryBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().nullish(),
  type: zod.enum(["text", "image"]).optional(),
  enabled: zod.boolean().optional(),
});

export const AdminUpdateCategoryResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string().nullish(),
  type: zod.enum(["text", "image"]),
  enabled: zod.boolean(),
  itemCount: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a category
 */
export const AdminDeleteCategoryParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminDeleteCategoryHeader = zod.object({
  "x-admin-password": zod.string(),
});

export const AdminDeleteCategoryResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List items in a category
 */
export const AdminListItemsParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminListItemsHeader = zod.object({
  "x-admin-password": zod.string(),
});

export const AdminListItemsResponseItem = zod.object({
  id: zod.number(),
  categoryId: zod.number(),
  itemText: zod.string(),
  imageUrl: zod.string().nullish(),
  createdAt: zod.string(),
});
export const AdminListItemsResponse = zod.array(AdminListItemsResponseItem);

/**
 * @summary Upload items via CSV/XLSX
 */
export const AdminUploadItemsParams = zod.object({
  id: zod.coerce.number(),
});

export const AdminUploadItemsHeader = zod.object({
  "x-admin-password": zod.string(),
});

export const AdminUploadItemsBody = zod.object({
  file: zod.instanceof(File),
});

export const AdminUploadItemsResponse = zod.object({
  imported: zod.number(),
  skipped: zod.number(),
  errors: zod.array(zod.string()),
  replaced: zod.boolean(),
});

/**
 * @summary Verify admin password
 */
export const AdminVerifyBody = zod.object({
  password: zod.string(),
});

export const AdminVerifyResponse = zod.object({
  valid: zod.boolean(),
});

/**
 * @summary Create a new room
 */
export const CreateRoomBody = zod.object({
  hostName: zod.string(),
  categoryId: zod.number().nullish(),
  turnDuration: zod.number().nullish(),
  roundCount: zod.number().nullish(),
});

/**
 * @summary Get room by code
 */
export const GetRoomParams = zod.object({
  code: zod.coerce.string(),
});

export const GetRoomResponse = zod.object({
  id: zod.number(),
  code: zod.string(),
  status: zod.enum(["waiting", "playing", "finished"]),
  categoryId: zod.number().nullish(),
  categoryName: zod.string().nullish(),
  turnDuration: zod.number(),
  roundCount: zod.number(),
  players: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      isHost: zod.boolean(),
      score: zod.number(),
      connected: zod.boolean(),
    }),
  ),
  createdAt: zod.string(),
});

/**
 * @summary Join a room
 */
export const JoinRoomParams = zod.object({
  code: zod.coerce.string(),
});

export const JoinRoomBody = zod.object({
  playerName: zod.string(),
});

export const JoinRoomResponse = zod.object({
  room: zod.object({
    id: zod.number(),
    code: zod.string(),
    status: zod.enum(["waiting", "playing", "finished"]),
    categoryId: zod.number().nullish(),
    categoryName: zod.string().nullish(),
    turnDuration: zod.number(),
    roundCount: zod.number(),
    players: zod.array(
      zod.object({
        id: zod.number(),
        name: zod.string(),
        isHost: zod.boolean(),
        score: zod.number(),
        connected: zod.boolean(),
      }),
    ),
    createdAt: zod.string(),
  }),
  playerId: zod.number(),
  playerName: zod.string(),
});