# lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryType } from "./categoryType";

export interface Category {
  id: number;
  name: string;
  description?: string | null;
  type: CategoryType;
  enabled: boolean;
  itemCount: number;
  createdAt: string;
}