# lib/api-zod/src/generated/types/categoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryItem {
  id: number;
  categoryId: number;
  itemText: string;
  imageUrl?: string | null;
  createdAt: string;
}