# lib/api-zod/src/generated/types/createCategoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCategoryBodyType } from "./createCategoryBodyType";

export interface CreateCategoryBody {
  name: string;
  description?: string | null;
  type: CreateCategoryBodyType;
  enabled?: boolean;
}