# lib/api-zod/src/generated/types/createRoomBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateRoomBody {
  hostName: string;
  categoryId?: number | null;
  turnDuration?: number | null;
  roundCount?: number | null;
}