# lib/api-zod/src/generated/types/joinRoomResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */
import type { RoomWithPlayers } from "./roomWithPlayers";

export interface JoinRoomResult {
  room: RoomWithPlayers;
  playerId: number;
  playerName: string;
}