# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  name: string;
  isHost: boolean;
  score: number;
  connected: boolean;
}