# lib/api-zod/src/generated/types/roomWithPlayers.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from "./player";
import type { RoomWithPlayersStatus } from "./roomWithPlayersStatus";

export interface RoomWithPlayers {
  id: number;
  code: string;
  status: RoomWithPlayersStatus;
  categoryId?: number | null;
  categoryName?: string | null;
  turnDuration: number;
  roundCount: number;
  players: Player[];
  createdAt: string;
}