# lib/api-zod/src/generated/types/updateCategoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Forehead Guess Game API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateCategoryBodyType } from "./updateCategoryBodyType";

export interface UpdateCategoryBody {
  name?: string;
  description?: string | null;
  type?: UpdateCategoryBodyType;
  enabled?: boolean;
}